Renovation phase 2
Added createNewOverlay function() and an overlayIndex reactive var.  overlayIndex is the index of the currently being edited overlay.  Multiple overlays can now be created for a doc and are stored in an array.  The other overlays (all but the one being edited) are used to create Stub components -- they display but are not affected by edit operations.