Positioning fixes
Overlay positions now saved as percentage of the window height (images always square for now) and  styled with "vh" units.  Overlays have an object to hold their data and Editor keeps an array of these.  Whenever an overlay is saved, it saves it's vh value.  Both Overlay and Stub expect the percentage value.  Whenever an overlay is mounted, it uses it's saved position data, but then responds to sizing events.